@@ -1,7 +1,7 @@
-const { Sequelize, DataTypes } = require('sequelize');
-const { sequelize } = require("../config/connection");
-const { Customer } = require('./customer');
-const { Product } = require('./products');
+import { DataTypes, Sequelize } from 'sequelize';
+import { sequelize } from "../config/connection.js";
+import { Customer } from './customer.js';
+import { Product } from './products.js';
 
 //Creating cart model
 const Cart = sequelize.define('Cart', {
@@ -47,4 +47,4 @@ Product.belongsToMany(Cart, { through: CartItem });
 Cart.sync();
 CartItem.sync();
 
-module.exports = { Cart, CartItem };
+export { Cart, CartItem };
